DTO 리팩토링 - BinaryContent, ReadStatus, UserStatus
