+++ src/main/java/com/sprint/mission/discodeit/controller/BinaryFileController.java
@@ -1,8 +1,7 @@
 package com.sprint.mission.discodeit.controller;
-import com.sprint.mission.discodeit.dto.BinaryContentService.BinaryContentResponseDto;
-import com.sprint.mission.discodeit.dto.BinaryContentService.BinaryContentResponseDtos;
+
+import com.sprint.mission.discodeit.dto.BinaryContentDto.*;
 import com.sprint.mission.discodeit.entity.BinaryContent;
-import com.sprint.mission.discodeit.mapper.BinaryContentMapper;
 import com.sprint.mission.discodeit.service.BinaryContentService;
 import io.swagger.v3.oas.annotations.tags.Tag;
 import lombok.RequiredArgsConstructor;
@@ -17,18 +16,21 @@ import java.util.UUID;
 @Tag(name = "BinaryFile", description = "Binary File 관련 API")
 public class BinaryFileController {
 
-    private final BinaryContentService binaryContentService;
+  private final BinaryContentService binaryContentService;
 
-    @GetMapping
-    public ResponseEntity<BinaryContentResponseDtos> getBinaryContents(@RequestParam("ids") List<UUID> binaryContentUUIDList
-    ) {
-        BinaryContentResponseDtos binaryContentResponseDtos = binaryContentService.findAllByIdIn(binaryContentUUIDList);
-        return ResponseEntity.ok(binaryContentResponseDtos);
-    }
+  @GetMapping
+  public ResponseEntity<BinaryContentResponseDtos> getBinaryContents(
+      @RequestParam("ids") List<UUID> binaryContentUUIDList
+  ) {
+    BinaryContentResponseDtos binaryContentResponseDtos = binaryContentService.findAllByIdIn(
+        binaryContentUUIDList);
+    return ResponseEntity.ok(binaryContentResponseDtos);
+  }
 
-    @GetMapping("/find")
-    public ResponseEntity<BinaryContent> getBinaryContent(@RequestParam("binaryContentId") UUID binaryContentId) {
-        BinaryContent binaryContent = binaryContentService.find(binaryContentId);
-        return ResponseEntity.ok(binaryContent);
-    }
+  @GetMapping("/find")
+  public ResponseEntity<BinaryContent> getBinaryContent(
+      @RequestParam("binaryContentId") UUID binaryContentId) {
+    BinaryContent binaryContent = binaryContentService.find(binaryContentId);
+    return ResponseEntity.ok(binaryContent);
+  }
 }
