+++ src/main/java/com/sprint/mission/discodeit/controller/ReadStatusController.java
@@ -1,13 +1,9 @@
 package com.sprint.mission.discodeit.controller;
-import com.sprint.mission.discodeit.dto.ReadStatusService.ReadStatusRequestDto;
-import com.sprint.mission.discodeit.dto.ReadStatusService.ReadStatusResponseDto;
-import com.sprint.mission.discodeit.dto.ReadStatusService.ReadStatusResponseDtos;
-import com.sprint.mission.discodeit.dto.ReadStatusService.UpdateReadStatusDto;
-import com.sprint.mission.discodeit.dto.UserStatusService.UpdateUserStatusDto;
+
+import com.sprint.mission.discodeit.dto.ReadStatusDto.*;
 import com.sprint.mission.discodeit.service.ReadStatusService;
 import io.swagger.v3.oas.annotations.tags.Tag;
 import lombok.RequiredArgsConstructor;
-import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
@@ -19,26 +15,28 @@ import java.util.UUID;
 @Tag(name = "ReadStatus", description = "ReadStatus 관련 API")
 public class ReadStatusController {
 
-    private final ReadStatusService readStatusService;
+  private final ReadStatusService readStatusService;
 
-    @PostMapping // 등록
-    public ResponseEntity<ReadStatusResponseDto> createReadStatus(@RequestBody ReadStatusRequestDto readStatusRequestDto
-    ) {
-        ReadStatusResponseDto readStatusResponseDto = readStatusService.create(readStatusRequestDto);
-        return ResponseEntity.ok().body(readStatusResponseDto);
-    }
+  @PostMapping // 등록
+  public ResponseEntity<ReadStatusResponseDto> createReadStatus(
+      @RequestBody ReadStatusRequestDto readStatusRequestDto
+  ) {
+    ReadStatusResponseDto readStatusResponseDto = readStatusService.create(readStatusRequestDto);
+    return ResponseEntity.ok().body(readStatusResponseDto);
+  }
 
-    @PatchMapping("/{readStatus-id}") // readStatus 업데이트
-    public ResponseEntity<String>  updateUserStatus(@PathVariable("readStatus-id") UUID readStatusId
-    ) {
-        readStatusService.update(readStatusId);
-        return ResponseEntity.ok().body("ReadStatus successfully updated");
-    }
+  @PatchMapping("/{readStatus-id}") // readStatus 업데이트
+  public ResponseEntity<String> updateUserStatus(@PathVariable("readStatus-id") UUID readStatusId
+  ) {
+    readStatusService.update(readStatusId);
+    return ResponseEntity.ok().body("ReadStatus successfully updated");
+  }
 
-    @GetMapping// 특정 사용자의 메시지 수신 정보 조회
-    public ResponseEntity<ReadStatusResponseDtos> getReadStatusByUserId(@RequestParam("user-Id") UUID userId
-    ) {
-        ReadStatusResponseDtos readStatusResponseDtos = readStatusService.findAllByUserId(userId);
-        return ResponseEntity.ok(readStatusResponseDtos);
-    }
+  @GetMapping// 특정 사용자의 메시지 수신 정보 조회
+  public ResponseEntity<ReadStatusResponseDtos> getReadStatusByUserId(
+      @RequestParam("user-Id") UUID userId
+  ) {
+    ReadStatusResponseDtos readStatusResponseDtos = readStatusService.findAllByUserId(userId);
+    return ResponseEntity.ok(readStatusResponseDtos);
+  }
 }
