--- src/main/java/com/sprint/mission/discodeit/dto/ReadStatusService/ReadStatusRequestDto.java
@@ -1,9 +1,0 @@
-package com.sprint.mission.discodeit.dto.ReadStatusService;
-
-import java.time.Instant;
-import java.util.UUID;
-
-public record ReadStatusRequestDto(
-        UUID userId,
-        UUID channelId
-) {}
