--- src/main/java/com/sprint/mission/discodeit/dto/ReadStatusService/UpdateReadStatusDto.java
@@ -1,8 +1,0 @@
-package com.sprint.mission.discodeit.dto.ReadStatusService;
-
-import java.time.Instant;
-import java.util.UUID;
-
-public record UpdateReadStatusDto(
-    Instant lastReadAt
-) {}
