--- src/main/java/com/sprint/mission/discodeit/dto/UserStatusService/UpdateUserStatusDto.java
@@ -1,7 +1,0 @@
-package com.sprint.mission.discodeit.dto.UserStatusService;
-import com.sprint.mission.discodeit.entity.UserStatus;
-import java.util.UUID;
-
-public record UpdateUserStatusDto(
-        UUID userStatusId
-) {}
