--- src/main/java/com/sprint/mission/discodeit/dto/UserStatusService/UserStatusResponseDtos.java
@@ -1,8 +1,0 @@
-package com.sprint.mission.discodeit.dto.UserStatusService;
-
-import java.util.ArrayList;
-import java.util.List;
-
-public record UserStatusResponseDtos(
-        List<UserStatusResponseDto> userStatusResponseDtos
-) {}
