+++ src/main/java/com/sprint/mission/discodeit/mapper/BinaryContentMapper.java
@@ -1,7 +1,6 @@
 package com.sprint.mission.discodeit.mapper;
-import com.sprint.mission.discodeit.dto.BinaryContentService.BinaryContentRequestDto;
-import com.sprint.mission.discodeit.dto.BinaryContentService.BinaryContentResponseDto;
-import com.sprint.mission.discodeit.dto.BinaryContentService.BinaryContentResponseDtos;
+
+import com.sprint.mission.discodeit.dto.BinaryContentDto.*;
 import com.sprint.mission.discodeit.entity.BinaryContent;
 import org.springframework.stereotype.Component;
 import org.springframework.web.multipart.MultipartFile;
@@ -13,41 +12,41 @@ import java.util.UUID;
 @Component
 public class BinaryContentMapper {
 
-    // Request
-    public BinaryContent toBinaryContent(UUID userId, UUID messageId, MultipartFile file) {
+  // Request
+  public BinaryContent toBinaryContent(UUID userId, UUID messageId, MultipartFile file) {
 
-        if (file == null) {
-            return null;
-        }
-        try {
-            return new BinaryContent(
-                    userId,
-                    messageId,
-                    file.getBytes(),
-                    file.getOriginalFilename(),
-                    file.getContentType()
-            );
-        } catch (IOException e){
-            throw new RuntimeException("Error reading file bytes");
-        }
+    if (file == null) {
+      return null;
     }
+    try {
+      return new BinaryContent(
+          userId,
+          messageId,
+          file.getBytes(),
+          file.getOriginalFilename(),
+          file.getContentType()
+      );
+    } catch (IOException e) {
+      throw new RuntimeException("Error reading file bytes");
+    }
+  }
 
-    // Response
-    public BinaryContentResponseDto toBinaryContentResponseDto(BinaryContent binaryContent) {
-        return new BinaryContentResponseDto(
-                binaryContent.getId(),
-                binaryContent.getUserId(),
-                binaryContent.getMessageId()
-        );
-    }
+  // Response
+  public BinaryContentResponseDto toBinaryContentResponseDto(BinaryContent binaryContent) {
+    return new BinaryContentResponseDto(
+        binaryContent.getId(),
+        binaryContent.getUserId(),
+        binaryContent.getMessageId()
+    );
+  }
 
-    // Response
-    public BinaryContentResponseDtos toBinaryContentResponseDtos(List<BinaryContent> binaryContents) {
-        return new BinaryContentResponseDtos(
-                binaryContents
-                        .stream()
-                        .map(this::toBinaryContentResponseDto)
-                        .toList()
-        );
-    }
+  // Response
+  public BinaryContentResponseDtos toBinaryContentResponseDtos(List<BinaryContent> binaryContents) {
+    return new BinaryContentResponseDtos(
+        binaryContents
+            .stream()
+            .map(this::toBinaryContentResponseDto)
+            .toList()
+    );
+  }
 }
